Track the requested state so we can return it when asked
The original USB relay code can query the GPIO lines to see which relays
are enabled. We only have one way communication to the power sockets so
they might be in a different state to that requested, but store the
state we think we should have so when asked we can return it, rather
than always returning that everything is off.